+++ src/Propulsion.Cosmos/ChangeFeedProcessor.fs
@@ -68,20 +68,20 @@ type ChangeFeedObserverFactory =
     static member FromFunction (f : unit -> #IChangeFeedObserver) =
         { new IChangeFeedObserverFactory with member __.CreateObserver () = f () :> _ }
 
-type CosmosCollectionId = { database: string; collection: string }
+type ContainerId = { database: string; container: string }
 
 //// Wraps the [Azure CosmosDb ChangeFeedProcessor library](https://github.com/Azure/azure-documentdb-changefeedprocessor-dotnet)
 type ChangeFeedProcessor =
     static member Start
-        (   log : ILogger, discovery: Equinox.Cosmos.Discovery, connectionPolicy : ConnectionPolicy, source : CosmosCollectionId,
-            /// The aux, non-partitioned collection holding the partition leases.
-            // Aux coll should always read from the write region to keep the number of write conflicts to minimum when the sdk
+        (   log : ILogger, discovery: Equinox.Cosmos.Discovery, connectionPolicy : ConnectionPolicy, source : ContainerId,
+            /// The aux, non-partitioned container holding the partition leases.
+            // Aux container should always read from the write region to keep the number of write conflicts to a minimum when the sdk
             // updates the leases. Since the non-write region(s) might lag behind due to us using non-strong consistency, during
             // failover we are likely to reprocess some messages, but that's okay since processing has to be idempotent in any case
-            aux : CosmosCollectionId,
+            aux : ContainerId,
             createObserver : unit -> IChangeFeedObserver,
             ?leaseOwnerId : string,
-            /// Used to specify an endpoint/account key for the aux collection, where that varies from that of the source collection. Default: use `discovery`
+            /// Used to specify an endpoint/account key for the aux container, where that varies from that of the source container. Default: use `discovery`
             ?auxDiscovery : Equinox.Cosmos.Discovery,
             /// Identifier to disambiguate multiple independent feed processor positions (akin to a 'consumer group')
             ?leasePrefix : string,
@@ -132,8 +132,8 @@ type ChangeFeedProcessor =
                 DocumentCollectionInfo(Uri = u, DatabaseName = d, CollectionName = c, MasterKey = k, ConnectionPolicy = connectionPolicy)
             ChangeFeedProcessorBuilder()
                 .WithHostName(leaseOwnerId)
-                .WithFeedCollection(mk discovery source.database source.collection)
-                .WithLeaseCollection(mk (defaultArg auxDiscovery discovery) aux.database aux.collection)
+                .WithFeedCollection(mk discovery source.database source.container)
+                .WithLeaseCollection(mk (defaultArg auxDiscovery discovery) aux.database aux.container)
                 .WithProcessorOptions(feedProcessorOptions)
         match reportLagAndAwaitNextEstimation with
         | None -> ()